% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{blood.ref.333cpgs}
\alias{blood.ref.333cpgs}
\title{Methylation reference for cell types in blood}
\format{An object of class `"matrix"` with 333 CpG sites
        as rows and cell types as columns}
\usage{
data(blood.ref.333cpgs)
}
\description{
Provides expected methylation levels for cell type
markers in blood for estimating cell type fractions
}
\examples{
data(blood.ref.333cpgs)
cell.types <- colnames(blood.ref.333cpgs)
}
\references{
Teschendorff et al. 2017, BMC Bioinformatics.
}
\keyword{datasets}
