% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime_comparison.r
\name{celldmc}
\alias{celldmc}
\title{Run CellDMC method}
\usage{
celldmc(X, W, Y)
}
\description{
methods vector should have these exact function names. Input should always
be X, W, Y as produced by the simulation framework
}
