% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_w_comparison.r
\name{compare_cell_fraction_estimates}
\alias{compare_cell_fraction_estimates}
\title{Compares cell fraction estimates from EpiDISH and TCA under several models}
\usage{
compare_cell_fraction_estimates(
  X,
  W.facs,
  C1 = NULL,
  C2 = NULL,
  verbose = TRUE
)
}
\arguments{
\item{X}{Matrix of methylation values with column names as samples and row names as sites}

\item{W.facs}{Matrix of FACS derived cell type fractions with column names as cell types
and row names as samples}

\item{C1}{Matrix of source-specific covariates for TCA with column names as features and
row names as samples}

\item{C2}{Matrix of mixture-level covariates for TCA with column names as features and
row names as samples}
}
\value{
A list of results. Slot performance.metrics contains a dataframe where each
        row indicates either the PCC or RMSE of a method for estimating a specific
        cell type fraction. Slot prop.estimates contains a list of fractions estimated
        by each method.
}
\description{
Compares cell fraction estimates from EpiDISH and TCA under several models
}
