% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime_comparison.r
\name{compare_runtimes}
\alias{compare_runtimes}
\title{Compare method runtimes across various input sizes}
\usage{
compare_runtimes(results_dir, plot_type)

compare_runtimes(results_dir, plot_type)
}
\arguments{
\item{results_dir}{Directory to save plot of results}

\item{plot_type}{Extension for saving plot graphics, such as pdf or png}

\item{sample.sizes}{Numeric vector for sample sizes to simulate.}

\item{number.sites}{Numeric vector for number of features/sites to simulate.}

\item{methods}{Character vector of method wrapper names to benchmark.
Examples of wrappers (tca.mle, tca.gmm, celldmc) are above. They
must have the same input arguments. Function names must
match to be called correctly.}

\item{method.names}{Pretty names of those listed in methods argument for plotting}

\item{number.replications}{Integer indicating how many replications of
each experiment to run. Same input is run for
each replicate. This is meant to account for
variability in runtime due to the system we are
running on.}

\item{number.cell.types}{Integer indicating number of cell types to simulate.}

\item{seed}{Integer to set seed for simulation to ensure replicability}
}
\value{
A list of dataframes. First dataframe with columns indicating the method,
        replicate number, sample size, number of sites, number of sources, and
        total runtime for each experiment. Second dataframe summarizes replicates.

A list of the runtime results and plot
}
\description{
Compare method runtimes across various input sizes

Compare runtimes of CellDMC and TCA (GMM and MLE modes)
}
