% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_mle_gmm_comparison.r
\name{compare_vars_mle}
\alias{compare_vars_mle}
\title{Compares sigma and mu esimtation of TCA with or without vars.mle option enabled}
\usage{
compare_vars_mle(
  sample.sizes,
  n.replications,
  n.sites,
  n.cell.types,
  n.c1.covs,
  n.c2.covs,
  tau,
  random_seed = 1000
)
}
\description{
Compares sigma and mu esimtation of TCA with or without vars.mle option enabled
}
