% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_w_comparison.r
\name{epic_qc}
\alias{epic_qc}
\title{preprocessing for BBC dataset}
\usage{
epic_qc(idat_dir, data_dir)
}
\arguments{
\item{idat_dir}{Directory storing IDAT files}

\item{data_dir}{Directory to store output beta values.}
}
\value{
Path to the beta value csv
}
\description{
preprocessing for BBC dataset
}
