% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_simulations.R
\name{parametric_simulations}
\alias{parametric_simulations}
\title{Replicate parametric simulations for power analyses}
\usage{
parametric_simulations(
  data_dir,
  results_dir,
  plot_dir,
  plot_type = "tiff",
  experiment_index = c(1, 2, 3, 4),
  num_cores = 1,
  m = 1000,
  m.true = 100,
  effect_sizes = 1:8,
  n = 500,
  num_sims = 50
)
}
\arguments{
\item{data_dir}{Path to directory to store data}

\item{results_dir}{Path to directory to store results}

\item{plot_dir}{Path to directory to store plots}

\item{plot_type}{Type of image to save plot}

\item{experiment_index}{An integer in 1, 2, 3, or 4. 1 specifies phenotype
affecting cell-type-specific methylation levels.
2 specifies the same as one but true positives
also consider predicted direction.
3 specifies the phenotype being affected by
cell-type-specific methylation levels. 4 is the
same as 1 with smaller sample size}

\item{num_cores}{Number of cores to use for analysis}

\item{m}{Number CpGs to simulate}

\item{m.true}{Number of true associations}

\item{effect_sizes}{Vector of effect sizes to simulate}

\item{n}{Sample size}

\item{num_sims}{Number of simulations to run}
}
\value{
None (TODO: Return a list of results)
}
\description{
Replicate parametric simulations for power analyses
}
