% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_w_comparison.r
\name{plot_results}
\alias{plot_results}
\title{Plot performance metrics for BBC and Koestler datasets}
\usage{
plot_results(
  bbc.results,
  koestler.results,
  bbc.sample.size,
  koestler.sample.size
)
}
\arguments{
\item{bbc.results}{List output by compare_cell_fraction_estimates() for BBC dataset}

\item{koestler.results}{List output by compare_cell_fraction_estimates() for Koestler dataset}

\item{bbc.sample.size}{Samples used in BBC dataset}

\item{koestler.sample.size}{Samples used in Koestler dataset}
}
\description{
Plot performance metrics for BBC and Koestler datasets
}
