% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_w_comparison.r
\name{prep_bbc_data}
\alias{prep_bbc_data}
\title{Downloads and processes raw BBC methylation data}
\usage{
prep_bbc_data(data_dir, bbc_pheno_path, bbc_facs_path)
}
\arguments{
\item{data_dir}{Directory to save bbc.RData}

\item{bbc_pheno_path}{Path to PhenoTypesBBC.Rd (provided by Jing et al.) which contains
covariates for the BBC dataset}

\item{bbc_facs_path}{Path to facsBBC.Rd (provided by Jing et al.) which contains 
FACS-based cell fractions for the BBC dataset}
}
\value{
The bbc object for analysis. Also saves an RData file to specified directory
}
\description{
Downloads and processes raw BBC methylation data
}
