% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking_analysis.R
\name{prep_smoking_data}
\alias{prep_smoking_data}
\title{Download and organize Liu et al. and Hannum et al. Data}
\usage{
prep_smoking_data(data_path, hannum_smk_status_path)
}
\arguments{
\item{data_path}{Directory for RData files that will store data}

\item{hannum_smk_status_path}{Path to hannum_smoking_status.txt
that must be downloaded separately.}
}
\value{
List of filenames for these RData files
}
\description{
Download and organize Liu et al. and Hannum et al. Data
}
