% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking_analysis.R
\name{run_smoking_analysis}
\alias{run_smoking_analysis}
\title{Smoking analysis of methylation data}
\usage{
run_smoking_analysis(X, smk, covs, W, smoking_cpgs, dataset.name)
}
\arguments{
\item{X}{Methylation values in matrix with sites as rows and samples as columns}

\item{smk}{Smoking status encoded as 0 (non-smoker), 1 (ex-smoker), or 2 (smoker) in
a matrix with rownames as samples and 1 column entitled \code{smoking}}

\item{covs}{Covariates in matrix format. Rownames are samples and column names are
covariate names}

\item{W}{Cell fractions in a matrix with cell types as column names and samples as
row names. Expects cell types: "Neutro","Eosino","Mono","CD4T","CD8T","B","NK"}

\item{smoking_cpgs}{List of CpG names to analyze for associations}

\item{dataset.name}{Name of dataset for naming output variables.}
}
\value{
A list of results. Each slot contains p-values output by each method
}
\description{
Performs cell-type-specific analysis of blood methylation data in the context of smoking.
Provides p-values estimated by CellDMC and TCA while correcting for given covariates
as well as ReFACTor components. Combines cell types into myeloid and lymphoid compartments
for analysis.
}
