% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime_comparison.r
\name{runtime_comparison}
\alias{runtime_comparison}
\title{Compare runtimes of CellDMC and TCA (GMM and MLE modes)}
\usage{
runtime_comparison(
  plot_dir,
  plot_type = "tiff",
  sample.sizes = c(50, 100, 200),
  number.sites = c(1000, 10000, 50000, 1e+05)
)
}
\arguments{
\item{plot_dir}{Directory to save plot of results}

\item{plot_type}{Extension for saving plot graphics, such as pdf or png}

\item{sample.sizes}{Vector of sample sizes to run}

\item{number.sites}{Vector of number of features to run}
}
\value{
A list of the runtime results and plot
}
\description{
Compare runtimes of CellDMC and TCA (GMM and MLE modes)
}
