% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime_comparison.r
\name{tca.gmm.reg}
\alias{tca.gmm.reg}
\title{Run fast TCA method with tcareg}
\usage{
tca.gmm.reg(X, W, Y)
}
\description{
methods vector should have these exact function names. Input should always
be X, W, Y as produced by the simulation framework
}
