% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime_comparison.r
\name{tca.mle}
\alias{tca.mle}
\title{Run original TCA method}
\usage{
tca.mle(X, W, Y)
}
\description{
methods vector should have these exact function names. Input should always
be X, W, Y as produced by the simulation framework
}
