% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_mle_gmm_comparison.r
\name{var_mle_gmm_comparison}
\alias{var_mle_gmm_comparison}
\title{Compare parameter estimation of TCA in MLE versus GMM Mode}
\usage{
var_mle_gmm_comparison(results_dir, plot_type)
}
\arguments{
\item{results_dir}{Directory to save plot of results}

\item{plot_type}{Extension for saving plot graphics, such as pdf or png}
}
\value{
A list of the results and plot
}
\description{
Compare parameter estimation of TCA in MLE versus GMM Mode
}
