% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime_comparison.r
\name{compare_runtimes}
\alias{compare_runtimes}
\title{Compare method runtimes across various input sizes}
\usage{
compare_runtimes(
  sample.sizes,
  number.sites,
  methods,
  method.names,
  number.replications,
  number.cell.types,
  random_seed
)
}
\arguments{
\item{sample.sizes}{Numeric vector for sample sizes to simulate.}

\item{number.sites}{Numeric vector for number of features/sites to simulate.}

\item{methods}{Character vector of method wrapper names to benchmark.
Examples of wrappers (tca.mle, tca.gmm, celldmc) are above. They
must have the same input arguments. Function names must
match to be called correctly.}

\item{method.names}{Pretty names of those listed in methods argument for plotting}

\item{number.replications}{Integer indicating how many replications of
each experiment to run. Same input is run for
each replicate. This is meant to account for
variability in runtime due to the system we are
running on.}

\item{number.cell.types}{Integer indicating number of cell types to simulate.}

\item{random_seed}{Integer to set seed for simulation to ensure replicability}
}
\value{
A list of dataframes. First dataframe with columns indicating the method,
        replicate number, sample size, number of sites, number of sources, and
        total runtime for each experiment. Second dataframe summarizes replicates.
}
\description{
Compare method runtimes across various input sizes
}
