% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_w_comparison.r
\name{epic_qc}
\alias{epic_qc}
\title{Preprocessing for BBC dataset with ENmix}
\usage{
epic_qc(idat_dir)
}
\arguments{
\item{idat_dir}{Directory storing IDAT files}
}
\value{
Matrix of beta values
}
\description{
Preprocessing for BBC dataset with ENmix
}
