% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_simulations.R
\name{parametric_simulations}
\alias{parametric_simulations}
\title{Replicate parametric simulations for power analyses}
\usage{
parametric_simulations(
  data_path,
  results_path,
  plots_path,
  experiment_index = c(1, 2, 3),
  num_cores = 1,
  image_format = "tiff"
)
}
\arguments{
\item{data_path}{Path to directory to store data}

\item{results_path}{Path to directory to store results}

\item{plots_path}{Path to directory to store plots}

\item{experiment_index}{An integer in 1, 2, or 3. 1 specifies phenotype
affecting cell-type-specific methylation levels.
2 specifies the same as one but true positives
also consider predicted direction.
3 specifies the phenotype being affected by
cell-type-specific methylation levels.}
}
\value{
None (TODO: Return a list of results)
}
\description{
Replicate parametric simulations for power analyses
}
