% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_w_comparison.r
\name{refit_w_comparison}
\alias{refit_w_comparison}
\title{Replicate cell type fraction analysis of Koestler et al. and BBC datasets}
\usage{
refit_w_comparison(
  bbc_pheno_path,
  bbc_facs_path,
  data_dir,
  plot_dir,
  plot_type = "tiff",
  n.sites = NULL
)
}
\arguments{
\item{bbc_pheno_path}{Path to PhenoTypesBBC.Rd (provided by Jing et al.) which contains
covariates for the BBC dataset}

\item{bbc_facs_path}{Path to facsBBC.Rd (provided by Jing et al.) which contains 
FACS-based cell fractions for the BBC dataset}

\item{data_dir}{Directory to store BBC data}

\item{plot_dir}{Directory to save plot of results}

\item{plot_type}{Extension for saving plot graphics, such as pdf or png}

\item{n.sites}{Number of sites to subset for quick testing purposes.}
}
\value{
A list of results. Slot \code{results} contains a list of proportions and metrics generated
        by each method on each dataset. Slot \code{plot} contains the plot object
}
\description{
Replicate cell type fraction analysis of Koestler et al. and BBC datasets
}
