% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking_analysis.R
\name{smoking_analysis}
\alias{smoking_analysis}
\title{Replicate smoking analysis of Liu et al. and Hannum et al. datasets}
\usage{
smoking_analysis(liu, hannum, results_dir, plot_type)
}
\arguments{
\item{liu}{List generated by get_liu_data() or provided in CellTypeSpecificMethylationData
package under the same name}

\item{hannum}{List generated by get_hannum_data() or provided in CellTypeSpecificMethylationData
package under the same name}

\item{results_dir}{Directory to save plot of results}

\item{plot_type}{Extension for saving plot graphics, such as pdf or png}
}
\value{
A list of results. Slot \code{results} contains a list of p-values generated by each method.
        Slot \code{plot} contains plot of results.
}
\description{
Replicate smoking analysis of Liu et al. and Hannum et al. datasets
}
